delete the complementar tracker version
I always use contextual tracking now, so complementar tracking seems
less interesting and i never feel the need to try it out. This file
does not seem to contain any particularly valuable logic.

If i will want to develop complementar tracking again in the future,
it will probably be contextual complementar tracking. That seems too
much complexity in interaction for the command line, so it is unlikely
to happen